Revert "LPA-2266: Fix for leading zeros bug and refactor"
