+++ src/Opg/Lpa/DataModel/Lpa/Elements/Dob.php
@@ -2,8 +2,10 @@
 
 namespace Opg\Lpa\DataModel\Lpa\Elements;
 
-use Opg\Lpa\DataModel\Common\Dob as BaseDob;
+use Opg\Lpa\DataModel\AbstractData;
+use Opg\Lpa\DataModel\Validator\Constraints as Assert;
 use Symfony\Component\Validator\Mapping\ClassMetadata;
+use DateTime;
 
 /**
  * Represents a date of birth.
@@ -11,12 +13,77 @@ use Symfony\Component\Validator\Mapping\ClassMetadata;
  * Class Dob
  * @package Opg\Lpa\DataModel\Lpa\Elements
  */
-class Dob extends BaseDob
+class Dob extends AbstractData
 {
+    /**
+     * A date of birth. The time component of the DateTime object should be ignored.
+     *
+     * @var DateTime
+     */
     protected $date;
 
     public static function loadValidatorMetadata(ClassMetadata $metadata)
     {
-        parent::loadValidatorMetadataCommon($metadata, "must-be-less-than-or-equal-to-today");
+        // As there is only 1 property, include NotBlank as there is no point this object existing without it.
+        $lessThanOrEqualToToday = new Assert\LessThanOrEqual([
+            'value' => new \DateTime('today')
+        ]);
+
+        $lessThanOrEqualToToday->message = "must-be-less-than-or-equal-to-today";
+
+        $metadata->addPropertyConstraints('date', [
+            new Assert\NotBlank,
+            new Assert\Custom\DateTimeUTC,
+            $lessThanOrEqualToToday,
+        ]);
+    }
+
+    /**
+     * @param string $property string Property name
+     * @param mixed $v mixed Value to map.
+     * @return mixed Mapped value.
+     */
+    protected function map($property, $v)
+    {
+        switch ($property) {
+            case 'date':
+                if ($v instanceof DateTime || is_null($v)) {
+                    return $v;
+                }
+
+                if (is_string($v)) {
+                    //  Assume the value is invalid until we have proved otherwise below
+                    $isValid = false;
+
+                    //  Split the array into components
+                    $dateArr = explode('-', $v);
+
+                    if (count($dateArr) == 3) {
+                        //  Truncate the day value to lose any time data and try to create a DateTime object
+                        $dateArr[2] = substr($dateArr[2], 0, 2);
+
+                        //  If required add any leading zeros to the day and month
+                        $dateArr[1] = str_pad($dateArr[1], 2, '0', STR_PAD_LEFT);
+                        $dateArr[2] = str_pad($dateArr[2], 2, '0', STR_PAD_LEFT);
+
+                        //  Format the string and date to the same format to ensure that it is valid
+                        $dateFormat = 'Y-m-d';
+                        $dateIn = implode('-', $dateArr);
+                        $date = DateTime::createFromFormat('Y-m-d', $dateIn);
+
+                        $isValid = ($date instanceof DateTime && strpos($dateIn, $date->format($dateFormat)) === 0);
+                    }
+
+                    if (!$isValid) {
+                        //  The date is invalid so return '0' instead of null
+                        //  This will allow the NotBlank validation to pass so we can display an appropriate date not valid message
+                        return '0';
+                    }
+                }
+
+                return new DateTime($v);
+        }
+
+        return parent::map($property, $v);
     }
 }
