+++ src/Opg/Lpa/DataModel/User/Dob.php
@@ -2,8 +2,10 @@
 
 namespace Opg\Lpa\DataModel\User;
 
-use Opg\Lpa\DataModel\Common\Dob as BaseDob;
+use Opg\Lpa\DataModel\AbstractData;
+use Opg\Lpa\DataModel\Validator\Constraints as Assert;
 use Symfony\Component\Validator\Mapping\ClassMetadata;
+use DateTime;
 use RuntimeException;
 
 /**
@@ -12,11 +14,23 @@ use RuntimeException;
  * Class Dob
  * @package Opg\Lpa\DataModel\Lpa\Elements
  */
-class Dob extends BaseDob
+class Dob extends AbstractData
 {
+    /**
+     * @var \DateTime A date of birth. The time component of the DateTime object should be ignored.
+     */
+    protected $date;
+
     public static function loadValidatorMetadata(ClassMetadata $metadata)
     {
-        self::loadValidatorMetadataCommon($metadata, null);
+        // As there is only 1 property, include NotBlank as there is no point this object existing without it.
+        $metadata->addPropertyConstraints('date', [
+            new Assert\NotBlank,
+            new Assert\Custom\DateTimeUTC,
+            new Assert\LessThanOrEqual([
+                'value' => new \DateTime('today')
+            ]),
+        ]);
     }
 
     /**
@@ -26,11 +40,23 @@ class Dob extends BaseDob
      */
     protected function map($property, $v)
     {
-        $mapped = parent::map($property, $v);
-        if ($mapped === null || $mapped === '0') {
-            throw new RuntimeException("Invalid date: $v. Date must exist and be in ISO-8601 format.");
+        switch ($property) {
+            case 'date':
+                if ($v instanceof DateTime || is_null($v)) {
+                    return $v;
+                }
+
+                if (is_string($v)) {
+                    $date = date_parse_from_format(DateTime::ISO8601, $v);
+
+                    if (!checkdate(@$date['month'], @$date['day'], @$date['year'])) {
+                        throw new RuntimeException("Invalid date: $v. Date must exist and be in ISO-8601 format.");
+                    }
+                }
+
+                return new DateTime($v);
         }
 
-        return $mapped;
+        return parent::map($property, $v);
     }
 }
